+++ libewf/libewf_file_entry.h
@@ -191,31 +191,31 @@ int libewf_file_entry_get_size(
 LIBEWF_EXTERN \
 int libewf_file_entry_get_creation_time(
      libewf_file_entry_t *file_entry,
-     int32_t *posix_time,
+     int64_t *posix_time,
      libcerror_error_t **error );
 
 LIBEWF_EXTERN \
 int libewf_file_entry_get_modification_time(
      libewf_file_entry_t *file_entry,
-     int32_t *posix_time,
+     int64_t *posix_time,
      libcerror_error_t **error );
 
 LIBEWF_EXTERN \
 int libewf_file_entry_get_access_time(
      libewf_file_entry_t *file_entry,
-     int32_t *posix_time,
+     int64_t *posix_time,
      libcerror_error_t **error );
 
 LIBEWF_EXTERN \
 int libewf_file_entry_get_entry_modification_time(
      libewf_file_entry_t *file_entry,
-     int32_t *posix_time,
+     int64_t *posix_time,
      libcerror_error_t **error );
 
 LIBEWF_EXTERN \
 int libewf_file_entry_get_deletion_time(
      libewf_file_entry_t *file_entry,
-     int32_t *posix_time,
+     int64_t *posix_time,
      libcerror_error_t **error );
 
 LIBEWF_EXTERN \
